[chore]:  파일 분리 및  코드 수정 #1
